+++ lib/src/rust/api/velopack.dart
@@ -8,30 +8,46 @@ import '../frb_generated.dart';
 import 'package:flutter_rust_bridge/flutter_rust_bridge_for_generated.dart';
 
 // These functions are ignored because they are not marked as `pub`: `check_and_download_updates`, `get_update_manager`
+// These types are ignored because they are neither used by any `pub` functions nor (for structs and enums) marked `#[frb(unignore)]`: `VelopackConfig`
 
-Future<void> initVelopack({required String url}) =>
-    VelopackRustLib.instance.api.crateApiVelopackInitVelopack(url: url);
+Future<void> initVelopack(
+        {required String url, String? channel, required bool allowDowngrade}) =>
+    VelopackRustLib.instance.api.crateApiVelopackInitVelopack(
+        url: url, channel: channel, allowDowngrade: allowDowngrade);
 
-Future<bool> isUpdateAvailable() =>
-    VelopackRustLib.instance.api.crateApiVelopackIsUpdateAvailable();
+Future<bool> isUpdateAvailable({String? channel, bool? allowDowngrade}) =>
+    VelopackRustLib.instance.api.crateApiVelopackIsUpdateAvailable(
+        channel: channel, allowDowngrade: allowDowngrade);
 
-Future<UpdateInfo?> getLatestUpdateInfo() =>
-    VelopackRustLib.instance.api.crateApiVelopackGetLatestUpdateInfo();
+Future<UpdateInfo?> getLatestUpdateInfo(
+        {String? channel, bool? allowDowngrade}) =>
+    VelopackRustLib.instance.api.crateApiVelopackGetLatestUpdateInfo(
+        channel: channel, allowDowngrade: allowDowngrade);
 
 Future<String> currentVersion() =>
     VelopackRustLib.instance.api.crateApiVelopackCurrentVersion();
 
-Stream<int> checkAndDownloadUpdatesWithProgress() =>
+Stream<int> checkAndDownloadUpdatesWithProgress(
+        {String? channel, bool? allowDowngrade}) =>
     VelopackRustLib.instance.api
-        .crateApiVelopackCheckAndDownloadUpdatesWithProgress();
+        .crateApiVelopackCheckAndDownloadUpdatesWithProgress(
+            channel: channel, allowDowngrade: allowDowngrade);
 
-Future<void> updateAndRestart() =>
-    VelopackRustLib.instance.api.crateApiVelopackUpdateAndRestart();
+Future<void> updateAndRestart({String? channel, bool? allowDowngrade}) =>
+    VelopackRustLib.instance.api.crateApiVelopackUpdateAndRestart(
+        channel: channel, allowDowngrade: allowDowngrade);
 
-Future<void> updateAndExit() =>
-    VelopackRustLib.instance.api.crateApiVelopackUpdateAndExit();
+Future<void> updateAndExit({String? channel, bool? allowDowngrade}) =>
+    VelopackRustLib.instance.api.crateApiVelopackUpdateAndExit(
+        channel: channel, allowDowngrade: allowDowngrade);
 
 Future<void> waitExitThenUpdate(
-        {required bool silent, required bool restart}) =>
-    VelopackRustLib.instance.api
-        .crateApiVelopackWaitExitThenUpdate(silent: silent, restart: restart);
+        {required bool silent,
+        required bool restart,
+        String? channel,
+        bool? allowDowngrade}) =>
+    VelopackRustLib.instance.api.crateApiVelopackWaitExitThenUpdate(
+        silent: silent,
+        restart: restart,
+        channel: channel,
+        allowDowngrade: allowDowngrade);