+++ rust/src/frb_generated.rs
@@ -71,6 +71,8 @@ fn wire__crate__api__velopack__check_and_download_updates_with_progress_impl(
                 <StreamSink<i16, flutter_rust_bridge::for_generated::SseCodec>>::sse_decode(
                     &mut deserializer,
                 );
+            let api_channel = <Option<String>>::sse_decode(&mut deserializer);
+            let api_allow_downgrade = <Option<bool>>::sse_decode(&mut deserializer);
             deserializer.end();
             move |context| {
                 transform_result_sse::<_, flutter_rust_bridge::for_generated::anyhow::Error>(
@@ -78,6 +80,8 @@ fn wire__crate__api__velopack__check_and_download_updates_with_progress_impl(
                         let output_ok =
                             crate::api::velopack::check_and_download_updates_with_progress(
                                 api_progress_sink,
+                                api_channel,
+                                api_allow_downgrade,
                             )?;
                         Ok(output_ok)
                     })(),
@@ -142,11 +146,16 @@ fn wire__crate__api__velopack__get_latest_update_info_impl(
             };
             let mut deserializer =
                 flutter_rust_bridge::for_generated::SseDeserializer::new(message);
+            let api_channel = <Option<String>>::sse_decode(&mut deserializer);
+            let api_allow_downgrade = <Option<bool>>::sse_decode(&mut deserializer);
             deserializer.end();
             move |context| {
                 transform_result_sse::<_, flutter_rust_bridge::for_generated::anyhow::Error>(
                     (move || {
-                        let output_ok = crate::api::velopack::get_latest_update_info()?;
+                        let output_ok = crate::api::velopack::get_latest_update_info(
+                            api_channel,
+                            api_allow_downgrade,
+                        )?;
                         Ok(output_ok)
                     })(),
                 )
@@ -211,11 +220,17 @@ fn wire__crate__api__velopack__init_velopack_impl(
             let mut deserializer =
                 flutter_rust_bridge::for_generated::SseDeserializer::new(message);
             let api_url = <String>::sse_decode(&mut deserializer);
+            let api_channel = <Option<String>>::sse_decode(&mut deserializer);
+            let api_allow_downgrade = <bool>::sse_decode(&mut deserializer);
             deserializer.end();
             move |context| {
                 transform_result_sse::<_, flutter_rust_bridge::for_generated::anyhow::Error>(
                     (move || {
-                        let output_ok = crate::api::velopack::init_velopack(api_url)?;
+                        let output_ok = crate::api::velopack::init_velopack(
+                            api_url,
+                            api_channel,
+                            api_allow_downgrade,
+                        )?;
                         Ok(output_ok)
                     })(),
                 )
@@ -245,11 +260,16 @@ fn wire__crate__api__velopack__is_update_available_impl(
             };
             let mut deserializer =
                 flutter_rust_bridge::for_generated::SseDeserializer::new(message);
+            let api_channel = <Option<String>>::sse_decode(&mut deserializer);
+            let api_allow_downgrade = <Option<bool>>::sse_decode(&mut deserializer);
             deserializer.end();
             move |context| {
                 transform_result_sse::<_, flutter_rust_bridge::for_generated::anyhow::Error>(
                     (move || {
-                        let output_ok = crate::api::velopack::is_update_available()?;
+                        let output_ok = crate::api::velopack::is_update_available(
+                            api_channel,
+                            api_allow_downgrade,
+                        )?;
                         Ok(output_ok)
                     })(),
                 )
@@ -279,11 +299,16 @@ fn wire__crate__api__velopack__update_and_exit_impl(
             };
             let mut deserializer =
                 flutter_rust_bridge::for_generated::SseDeserializer::new(message);
+            let api_channel = <Option<String>>::sse_decode(&mut deserializer);
+            let api_allow_downgrade = <Option<bool>>::sse_decode(&mut deserializer);
             deserializer.end();
             move |context| {
                 transform_result_sse::<_, flutter_rust_bridge::for_generated::anyhow::Error>(
                     (move || {
-                        let output_ok = crate::api::velopack::update_and_exit()?;
+                        let output_ok = crate::api::velopack::update_and_exit(
+                            api_channel,
+                            api_allow_downgrade,
+                        )?;
                         Ok(output_ok)
                     })(),
                 )
@@ -313,11 +338,16 @@ fn wire__crate__api__velopack__update_and_restart_impl(
             };
             let mut deserializer =
                 flutter_rust_bridge::for_generated::SseDeserializer::new(message);
+            let api_channel = <Option<String>>::sse_decode(&mut deserializer);
+            let api_allow_downgrade = <Option<bool>>::sse_decode(&mut deserializer);
             deserializer.end();
             move |context| {
                 transform_result_sse::<_, flutter_rust_bridge::for_generated::anyhow::Error>(
                     (move || {
-                        let output_ok = crate::api::velopack::update_and_restart()?;
+                        let output_ok = crate::api::velopack::update_and_restart(
+                            api_channel,
+                            api_allow_downgrade,
+                        )?;
                         Ok(output_ok)
                     })(),
                 )
@@ -349,12 +379,18 @@ fn wire__crate__api__velopack__wait_exit_then_update_impl(
                 flutter_rust_bridge::for_generated::SseDeserializer::new(message);
             let api_silent = <bool>::sse_decode(&mut deserializer);
             let api_restart = <bool>::sse_decode(&mut deserializer);
+            let api_channel = <Option<String>>::sse_decode(&mut deserializer);
+            let api_allow_downgrade = <Option<bool>>::sse_decode(&mut deserializer);
             deserializer.end();
             move |context| {
                 transform_result_sse::<_, flutter_rust_bridge::for_generated::anyhow::Error>(
                     (move || {
-                        let output_ok =
-                            crate::api::velopack::wait_exit_then_update(api_silent, api_restart)?;
+                        let output_ok = crate::api::velopack::wait_exit_then_update(
+                            api_silent,
+                            api_restart,
+                            api_channel,
+                            api_allow_downgrade,
+                        )?;
                         Ok(output_ok)
                     })(),
                 )
@@ -451,6 +487,28 @@ impl SseDecode for Vec<crate::core::dart_types::VelopackAsset> {
             ));
         }
         return ans_;
+    }
+}
+
+impl SseDecode for Option<String> {
+    // Codec=Sse (Serialization based), see doc to use other codecs
+    fn sse_decode(deserializer: &mut flutter_rust_bridge::for_generated::SseDeserializer) -> Self {
+        if (<bool>::sse_decode(deserializer)) {
+            return Some(<String>::sse_decode(deserializer));
+        } else {
+            return None;
+        }
+    }
+}
+
+impl SseDecode for Option<bool> {
+    // Codec=Sse (Serialization based), see doc to use other codecs
+    fn sse_decode(deserializer: &mut flutter_rust_bridge::for_generated::SseDeserializer) -> Self {
+        if (<bool>::sse_decode(deserializer)) {
+            return Some(<bool>::sse_decode(deserializer));
+        } else {
+            return None;
+        }
     }
 }
 
@@ -707,6 +765,26 @@ impl SseEncode for Vec<crate::core::dart_types::VelopackAsset> {
         <i32>::sse_encode(self.len() as _, serializer);
         for item in self {
             <crate::core::dart_types::VelopackAsset>::sse_encode(item, serializer);
+        }
+    }
+}
+
+impl SseEncode for Option<String> {
+    // Codec=Sse (Serialization based), see doc to use other codecs
+    fn sse_encode(self, serializer: &mut flutter_rust_bridge::for_generated::SseSerializer) {
+        <bool>::sse_encode(self.is_some(), serializer);
+        if let Some(value) = self {
+            <String>::sse_encode(value, serializer);
+        }
+    }
+}
+
+impl SseEncode for Option<bool> {
+    // Codec=Sse (Serialization based), see doc to use other codecs
+    fn sse_encode(self, serializer: &mut flutter_rust_bridge::for_generated::SseSerializer) {
+        <bool>::sse_encode(self.is_some(), serializer);
+        if let Some(value) = self {
+            <bool>::sse_encode(value, serializer);
         }
     }
 }